+++ rubato/utils/defaults.py
@@ -69,14 +69,14 @@ class Defaults:
         # The text to display.
         "justify": "left",  # . . . . . . . . . . . . . . . . . . . str
         # The justification of the text. (left, center, right)
-        "align": Vector(),  # . . . . . . . . . . . . . . . . . . . str
-        # The alignment of the text. The zero vector means it is centered.
+        "anchor": Vector(),  # . . . . . . . . . . . . . . . . . . . str
+        # The anchor position of the text. The zero vector means it is centered.
         # x component is whether to shift left, none, or right (-1, 0, 1)
         # y component is whether to shift top, none, or bottom (-1, 0, 1)
         "width": -1,  # . . . . . . . . . . . . . . . . . . . . . . int
         # The maximum width of the text. Will automatically wrap the text.
         "font": None,  #. . . . . . . . . . . . . . . . . . . . . . Font
-        # The font object to use. Defaults to an object using the Fredoka font.
+        # The font object to use. Defaults to the default Font object (Font()).
     }
     # [/text]
     # [font]